pyproject.toml
[project]
name = "audio-transkript"
version = "0.1.0"
description = "Lokales macOS Menu-Bar-Tool für Sprach-Transkription und Screenshot-OCR"
requires-python = ">=3.10"
dependencies = [
    "rumps>=0.4.0",
    "pyobjc-framework-Cocoa",
    "pyobjc-framework-Quartz",
    "pyobjc-framework-Vision",
    "sounddevice",
    "numpy",
    "mlx-whisper",
    "pynput",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"
